@ambv 2023-03-03
When looking at the PR without whitespace changes, it's not that complicated. All tests pass, no performance regression.

I think this should go in as is. This will help us optimize, whether with Cython or with mypyc, as the result is a plain old `Phase` class, which I like 👍🏻